Fixes document missing in Node env fix #594
